feat(CGU): add CGU
feat(ci): From GKE to Heroku

We have decided to abandon the deployment on GKE in favor of Heroku for pricing reasons.

From now on, the deployment will be managed by Heroku from the code contained on main.

CI must pass for a version to be deployed

fix(ci): Sonar Quality and Semantic Release

Fixed 3 security hotspot to allow Sonar to give us an A rate

Added semantic release

Signed-off-by: Robin JOSEPH <rjrdutg@gmail.com>